+++ libearth/parser/autodiscovery.py
@@ -15,13 +15,14 @@ try:
 except ImportError:
     import urllib.parse as urlparse
 
-from libearth.compat import text
-from libearth.parser.heuristic import get_format
-from libearth.parser import atom, rss2
+from ..compat import text
+from ..compat.etree import fromstring
+from .atom import parse_atom
+from .rss2 import parse_rss
 
 
 __all__ = ('ATOM_TYPE', 'RSS_TYPE', 'TYPE_TABLE', 'AutoDiscovery', 'FeedLink',
-           'FeedUrlNotFoundError', 'autodiscovery')
+           'FeedUrlNotFoundError', 'autodiscovery', 'get_format')
 
 
 #: (:class:`str`) The MIME type of RSS 2.0 format.
@@ -31,7 +32,7 @@ RSS_TYPE = 'application/rss+xml'
 ATOM_TYPE = 'application/atom+xml'
 
 #: (:class:`collections.Mapping`) The mapping table of feed types
-TYPE_TABLE = {atom.parse_atom: ATOM_TYPE, rss2.parse_rss: RSS_TYPE}
+TYPE_TABLE = {parse_atom: ATOM_TYPE, parse_rss: RSS_TYPE}
 
 #: Namedtuple which is a pair of ``type` and ``url``
 FeedLink = collections.namedtuple('FeedLink', 'type url')
@@ -118,3 +119,29 @@ class FeedUrlNotFoundError(Exception):
 
     def __init__(self, msg):
         self.msg = msg
+
+
+def get_format(document):
+    """Guess the syndication format of an arbitrary ``document``.
+
+    :param document: document string to guess
+    :type document: :class:`str`
+    :returns: the function possible to parse the given ``document``
+    :rtype: :class:`collections.Callable`
+
+    .. versionchanged:: 0.2.0
+       The function was in :mod:`libearth.parser.heuristic` module (which is
+       removed now) before 0.2.0, but now it's moved to
+       :mod:`libearth.parser.autodiscovery`.
+
+    """
+    try:
+        root = fromstring(document)
+    except Exception:
+        return None
+    if root.tag == '{http://www.w3.org/2005/Atom}feed':
+        return parse_atom
+    elif root.tag == 'rss':
+        return parse_rss
+    else:
+        return None